Added alternate button inputs for change weapon and changed all hp stats to floats

+++ FPSProject/PP2FPS/Assets/Scripts/IDamage.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public interface IDamage
 {
     // create the take damage function for interface
-    void takeDamage(int amount);
+    void takeDamage(float amount);
 }
